+++ changelog.md
@@ -2,7 +2,9 @@
 
 0.6 - TBD
 
-- 
+- Added `scan5(t)` code to scan for open NetBIOS ports, as possible responder targets.
+- Added `scan6(t)` code to scan for LLMNR resolutions, as possible respodner targets.
+- Added language for scan 5 and 6 in `README` and `oB-recon-configure.py`
 
 0.5 - 7/25/18